refactoring single_use_links_controller and single_use_links_viewer_controller into behavior modules
